+++ Changes.md
@@ -1,4 +1,14 @@
 
+
+### 2.0.0 - 2022-03-29
+
+- bump redis major version 3 -> 4
+- API change, callbacks replaced by promises
+- config.ini
+    - [server] -> [socket] 
+    - opts.db -> opts.database (to match upstream)
+
+
 ### 1.0.13 - 2021-10-14
 
 - switch CI from Travis to GitHub Actions